timezone_offset --> double precision

+++ src/objects/data_structures.f90
@@ -170,7 +170,7 @@ module data_structures
         integer, dimension(:), allocatable :: selected_level ! to just use a specific vertical level from each file (e.g. a pressure level)
                                                              ! the array is to provide one level for each variable
         integer :: interpolation_method
-        integer :: timezone_offset
+        double precision :: timezone_offset
 
         integer, dimension(:), allocatable :: time_indices   ! specific time indices to average over (e.g. multiple hours in a daily file)
         ! tranformation to apply to each atmophseric variable
